+++ src/Chunkyard/Core/IBlobWriter.cs
@@ -1,0 +1,16 @@
+﻿using System.IO;
+
+namespace Chunkyard.Core
+{
+    /// <summary>
+    /// Defines a contract to find, write and update Blobs.
+    /// </summary>
+    public interface IBlobWriter
+    {
+        Blob? FindBlob(string blobName);
+
+        Stream OpenWrite(string blobName);
+
+        void UpdateBlobMetadata(Blob blob);
+    }
+}